interpretation: co-locate unit tests with the functions they test
Per Nico's #883 review: the single 1822-line `interpretation/tests.rs` was too
big. Split it so each function's unit tests live in a `#[cfg(test)] mod tests`
in the same file as the function under test:
- prompt.rs — build_interpretation_input / system prompt / few-shot (5 tests)
- parse.rs  — parse_interpretation_response + retry_interpretation_parse (13)
- graph.rs  — plan_interpretation_ops(_with_context), relations, apply,
              existing_instance_context (15)
- run.rs    — strip_noop_updates, gather_transcript_sparql,
              ensure_interpretation_task (4)
- dedup.rs  — filter_already_present(_semantic/_with_strategy), cosine,
              semantic_dedup_pure (8)

Shared fixtures (existing_map/existing_ids/existing_by_identity/no_existing,
prop_values, nth_create_base/addlinks_for/targets_of, decoded_targets,
apply_one, proposal, seed_message) moved into the existing `#![cfg(test)]`
`interpretation_test_support` module as `pub(crate)`, so every per-file test
module shares one copy. `tests.rs` and its `mod tests;` are gone.

Pure move — no test body/assertion changes. All 45 interpretation tests pass
(--test-threads=1); no new warnings.
