+++ rust-executor/src/perspectives/interpretation/dedup.rs
@@ -247,3 +247,315 @@ pub async fn filter_already_present_with_strategy(
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::perspectives::interpretation::*;
+    use crate::perspectives::interpretation_test_support::*;
+    use std::collections::{BTreeMap, HashMap};
+
+    #[test]
+    fn filter_already_present_drops_known_titles() {
+        // Tasks declare `title` as their identity. Four Tasks proposed: one
+        // duplicates an existing title (case-insensitive), one duplicates it under
+        // whitespace normalization, one is new, and a same-title item of a
+        // DIFFERENT class (which has no identity here) is untouched (dedup is per
+        // class, and only for classes with a declared identity).
+        let proposed = parse_interpretation_response(
+            r#"[
+              {"class":"Task","title":"Ship the MVP"},
+              {"class":"Task","title":"  ship   the   mvp  "},
+              {"class":"Task","title":"Write the docs"},
+              {"class":"Belief","title":"ship the mvp"}
+            ]"#,
+        )
+        .unwrap();
+        let existing = existing_by_identity(&[("Task", "ship the MVP")]);
+        // Only Task declares `title` as its identity; Belief has none ⇒ no dedup.
+        let mut identity_props = HashMap::new();
+        identity_props.insert("Task".to_string(), "title".to_string());
+
+        let kept = filter_already_present(proposed, &existing, &identity_props);
+        let kept_titles: Vec<&str> = kept
+            .iter()
+            .filter_map(|i| i.props.get("title").and_then(|v| v.as_str()))
+            .collect();
+        assert!(
+            !kept_titles.contains(&"Ship the MVP"),
+            "existing Task title must be dropped (case-insensitive); got {kept_titles:?}"
+        );
+        assert!(
+            !kept_titles.contains(&"  ship   the   mvp  "),
+            "whitespace-normalized duplicate Task title must be dropped; got {kept_titles:?}"
+        );
+        assert!(
+            kept_titles.contains(&"Write the docs"),
+            "new Task must survive"
+        );
+        assert!(
+        kept_titles.contains(&"ship the mvp"),
+        "same title on a class with no declared identity must NOT be dropped; got {kept_titles:?}"
+    );
+    }
+
+    #[test]
+    fn cosine_similarity_orthogonal_and_parallel() {
+        // Sanity floor for the semantic dedup: parallel vectors → 1, opposite → -1,
+        // orthogonal → 0, zero → 0 (not NaN, so threshold comparisons still work
+        // without a special case in the filter).
+        let a = vec![1.0f32, 0.0, 0.0];
+        let b = vec![1.0f32, 0.0, 0.0];
+        assert!((cosine_similarity(&a, &b) - 1.0).abs() < 1e-6);
+        let c = vec![-1.0f32, 0.0, 0.0];
+        assert!((cosine_similarity(&a, &c) - (-1.0)).abs() < 1e-6);
+        let d = vec![0.0f32, 1.0, 0.0];
+        assert!(cosine_similarity(&a, &d).abs() < 1e-6);
+        let z = vec![0.0f32; 3];
+        assert_eq!(cosine_similarity(&a, &z), 0.0);
+    }
+
+    #[test]
+    fn semantic_dedup_pure_drops_near_duplicate_keeps_distinct() {
+        // The core invariant of the semantic dedup: a proposal whose embedding is
+        // near-parallel to an existing one (in this class) is dropped; one that
+        // isn't survives. Order of the survivors is preserved so relation
+        // `new:<Class>:<n>` ordinals still line up downstream. This test uses
+        // hand-crafted vectors so it exercises the filter without an HTTP round
+        // trip — the stub "embedder" is just the vectors we pass in.
+        let proposed = parse_interpretation_response(
+            r#"[
+              {"class":"Task","title":"Ship the MVP"},
+              {"class":"Task","title":"Ship MVP"},
+              {"class":"Task","title":"Write the docs"}
+            ]"#,
+        )
+        .unwrap();
+        // Existing vector for the only existing "Ship the MVP" task.
+        let mut existing_vecs: HashMap<String, Vec<Vec<f32>>> = HashMap::new();
+        existing_vecs.insert("Task".to_string(), vec![vec![1.0, 0.0, 0.0]]);
+        // Proposed vectors keyed by their index in `proposed`.
+        let mut proposed_vecs: HashMap<String, Vec<(usize, Vec<f32>)>> = HashMap::new();
+        proposed_vecs.insert(
+            "Task".to_string(),
+            vec![
+                (0, vec![1.0, 0.0, 0.0]),  // identical → sim 1.0, dropped
+                (1, vec![0.98, 0.2, 0.0]), // very close → sim ≈ 0.98, dropped
+                (2, vec![0.0, 1.0, 0.0]),  // orthogonal → sim 0, kept
+            ],
+        );
+        let kept = semantic_dedup_pure(proposed, &existing_vecs, &proposed_vecs, 0.85);
+        let kept_titles: Vec<&str> = kept
+            .iter()
+            .filter_map(|i| i.props.get("title").and_then(|v| v.as_str()))
+            .collect();
+        assert_eq!(
+            kept_titles,
+            vec!["Write the docs"],
+            "only the semantically distinct proposal survives; got {kept_titles:?}"
+        );
+    }
+
+    #[test]
+    fn semantic_dedup_pure_preserves_order_and_upsert() {
+        // Mirrors `filter_already_present_keeps_upserts_and_preserves_order` for
+        // the semantic path: proposals with an `id` are treated by the caller
+        // (`filter_already_present_semantic`) as upsert targets and never appear
+        // in `proposed_vecs`, so `semantic_dedup_pure` never drops them.
+        // Distinct titles surrounding a near-dup keep their relative order.
+        let proposed = parse_interpretation_response(
+            r#"[
+              {"class":"Task","title":"Alpha"},
+              {"class":"Task","id":"soa://existing/task/1","title":"Ship the MVP"},
+              {"class":"Task","title":"Ship MVP"},
+              {"class":"Task","title":"Omega"}
+            ]"#,
+        )
+        .unwrap();
+        let mut existing_vecs: HashMap<String, Vec<Vec<f32>>> = HashMap::new();
+        existing_vecs.insert("Task".to_string(), vec![vec![1.0, 0.0, 0.0]]);
+        // Only indices 0, 2, 3 are subject to dedup (index 1 has an `id`, so the
+        // outer caller skips it — mirror that here by NOT putting it in the map).
+        let mut proposed_vecs: HashMap<String, Vec<(usize, Vec<f32>)>> = HashMap::new();
+        proposed_vecs.insert(
+            "Task".to_string(),
+            vec![
+                (0, vec![0.0, 1.0, 0.0]),  // Alpha — orthogonal, kept
+                (2, vec![0.98, 0.2, 0.0]), // near-dup of existing, dropped
+                (3, vec![0.0, 0.0, 1.0]),  // Omega — orthogonal, kept
+            ],
+        );
+        let kept = semantic_dedup_pure(proposed, &existing_vecs, &proposed_vecs, 0.85);
+        let kept_titles: Vec<&str> = kept
+            .iter()
+            .filter_map(|i| i.props.get("title").and_then(|v| v.as_str()))
+            .collect();
+        assert_eq!(
+            kept_titles,
+            vec!["Alpha", "Ship the MVP", "Omega"],
+            "upsert not touched, near-dup dropped, order preserved"
+        );
+        assert_eq!(
+            kept[1].id.as_deref(),
+            Some("soa://existing/task/1"),
+            "surviving Ship the MVP is the id-carrying upsert"
+        );
+    }
+
+    #[tokio::test]
+    async fn dispatcher_normalized_string_matches_direct_call() {
+        // `filter_already_present_with_strategy(..., NormalizedString)` must
+        // behave identically to calling `filter_already_present` directly — the
+        // default path stays byte-for-byte compatible with existing callers.
+        let proposed = parse_interpretation_response(
+            r#"[
+              {"class":"Task","title":"Ship the MVP"},
+              {"class":"Task","title":"  ship   the   mvp  "},
+              {"class":"Task","title":"Write the docs"}
+            ]"#,
+        )
+        .unwrap();
+        let existing = existing_by_identity(&[("Task", "ship the MVP")]);
+        let mut identity_props = HashMap::new();
+        identity_props.insert("Task".to_string(), "title".to_string());
+
+        let via_dispatcher = filter_already_present_with_strategy(
+            proposed.clone(),
+            &existing,
+            &identity_props,
+            &DedupStrategy::default(),
+        )
+        .await
+        .unwrap();
+        let via_direct = filter_already_present(proposed, &existing, &identity_props);
+        let d_titles: Vec<&str> = via_dispatcher
+            .iter()
+            .filter_map(|i| i.props.get("title").and_then(|v| v.as_str()))
+            .collect();
+        let s_titles: Vec<&str> = via_direct
+            .iter()
+            .filter_map(|i| i.props.get("title").and_then(|v| v.as_str()))
+            .collect();
+        assert_eq!(d_titles, s_titles);
+        assert_eq!(d_titles, vec!["Write the docs"]);
+    }
+
+    #[test]
+    fn filter_already_present_keeps_upserts_and_preserves_order() {
+        // An `id`-carrying proposal is an explicit upsert target: its title
+        // deliberately matches an existing one, so dedup must never drop it. And
+        // filtering happens IN PLACE — the surviving order is the LLM's emission
+        // order, which is what `new:<Class>:<n>` ordinals count against. Both
+        // properties together are what make relation ordinals resolve correctly
+        // after a dedup pass.
+        let proposed = parse_interpretation_response(
+            r#"[
+              {"class":"Task","title":"Alpha"},
+              {"class":"Task","id":"soa://existing/task/1","title":"Ship the MVP"},
+              {"class":"Task","title":"Ship the MVP"},
+              {"class":"Task","title":"Omega"}
+            ]"#,
+        )
+        .unwrap();
+        // The existing instance carries both its id (the trusted upsert target) and
+        // its title (the dedup identity) — one row in the single-source map.
+        let existing = existing_map(vec![InstanceContext {
+            id: "soa://existing/task/1".to_string(),
+            title: "Ship the MVP".to_string(),
+            class: "Task".to_string(),
+            properties: BTreeMap::new(),
+        }]);
+        let mut identity_props = HashMap::new();
+        identity_props.insert("Task".to_string(), "title".to_string());
+
+        let kept = filter_already_present(proposed, &existing, &identity_props);
+        let kept_titles: Vec<&str> = kept
+            .iter()
+            .filter_map(|i| i.props.get("title").and_then(|v| v.as_str()))
+            .collect();
+        assert_eq!(
+            kept_titles,
+            vec!["Alpha", "Ship the MVP", "Omega"],
+            "upsert survives, plain duplicate is dropped, order preserved"
+        );
+        assert_eq!(
+            kept[1].id.as_deref(),
+            Some("soa://existing/task/1"),
+            "the surviving 'Ship the MVP' must be the id-carrying upsert"
+        );
+    }
+
+    #[test]
+    fn filter_already_present_dedups_hallucinated_id_but_keeps_trusted() {
+        // A proposal may carry an `id` the model invented. The planner only routes
+        // to Update for ids the graph actually holds (`known_existing_ids`); an
+        // untrusted id goes to Create — so it must still be dedup-checked, or a
+        // made-up id + duplicate identity mints a duplicate node. A *trusted* id is
+        // a real upsert target and bypasses dedup.
+        let proposed = parse_interpretation_response(
+            r#"[
+              {"class":"Task","id":"soa://hallucinated/999","title":"Ship the MVP"}
+            ]"#,
+        )
+        .unwrap();
+        let mut identity_props = HashMap::new();
+        identity_props.insert("Task".to_string(), "title".to_string());
+
+        // Untrusted id → the graph does not hold `soa://hallucinated/999`, only a
+        // same-title Task under a different id, so the proposal is deduped away.
+        let existing_untrusted = existing_by_identity(&[("Task", "Ship the MVP")]);
+        let dropped =
+            filter_already_present(proposed.clone(), &existing_untrusted, &identity_props);
+        assert!(
+            dropped.is_empty(),
+            "hallucinated id + duplicate identity must be deduped, not minted; got {dropped:#?}"
+        );
+
+        // Same proposal, but now the graph actually holds that id (title matches) →
+        // it is a real upsert target and bypasses dedup.
+        let existing_trusted = existing_map(vec![InstanceContext {
+            id: "soa://hallucinated/999".to_string(),
+            title: "Ship the MVP".to_string(),
+            class: "Task".to_string(),
+            properties: BTreeMap::new(),
+        }]);
+        let kept = filter_already_present(proposed, &existing_trusted, &identity_props);
+        assert_eq!(kept.len(), 1, "a trusted id bypasses dedup; got {kept:#?}");
+    }
+
+    #[test]
+    fn filter_already_present_dedupes_within_same_response() {
+        // The LLM sometimes emits the same (class, identity) twice in one response
+        // (verbatim, or under whitespace/case variation). Without intra-response
+        // dedup those slip past `filter_already_present` because the pre-existing
+        // `known` set does not yet contain them — and `run_interpretation` then
+        // mints two subjects for the same identity. Fix: accumulate accepted
+        // identities as we scan the response, dropping later same-key proposals
+        // exactly like already-persisted ones.
+        let proposed = parse_interpretation_response(
+            r#"[
+              {"class":"Task","title":"Ship the MVP"},
+              {"class":"Task","title":"  SHIP  the  mvp  "},
+              {"class":"Task","title":"Ship the MVP"},
+              {"class":"Task","title":"Write the docs"}
+            ]"#,
+        )
+        .unwrap();
+        let existing = ExistingInstances::new(); // graph empty
+        let mut identity_props = HashMap::new();
+        identity_props.insert("Task".to_string(), "title".to_string());
+
+        let kept = filter_already_present(proposed, &existing, &identity_props);
+        let kept_titles: Vec<&str> = kept
+            .iter()
+            .filter_map(|i| i.props.get("title").and_then(|v| v.as_str()))
+            .collect();
+
+        // First occurrence wins; every subsequent normalized-equal proposal drops.
+        assert_eq!(
+        kept_titles,
+        vec!["Ship the MVP", "Write the docs"],
+        "intra-response duplicates must be dropped after the first occurrence; got {kept_titles:?}"
+    );
+    }
+}
